pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "voice-rag"
version = "0.1.0"
description = "Provider-agnostic voice RAG pipeline. Plug in your voice provider, LLM, vector store, and document parsers."
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "GT", email = "gt@gaurangtorvekar.com" },
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "openai>=1.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.2.1",
    "qdrant-client>=1.10.0",
    "fastembed>=0.6.0",
    "httpx>=0.27.0",
    "click>=8.0.0",
    "python-multipart>=0.0.6",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
anthropic = ["anthropic>=0.40.0"]
gemini = ["google-genai>=1.0.0"]
deepgram = ["deepgram-sdk>=3.0.0"]
elevenlabs = ["elevenlabs>=1.0.0"]
pdf = ["pymupdf>=1.24.0"]
docx = ["python-docx>=1.0.0"]
all = [
    "voice-rag[anthropic,gemini,deepgram,elevenlabs,pdf,docx]",
]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "voice-rag[all]",
]

[project.scripts]
voice-rag = "voice_rag.cli.__main__:cli"

[project.urls]
Homepage = "https://github.com/kytona/voice-rag"
Repository = "https://github.com/kytona/voice-rag"

[tool.hatch.build.targets.wheel]
packages = ["voice_rag"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
